fix: vercel function appId
